Agent/requirements.txt



langchain
wikipedia
langchain-community
langchain-core
langchain-google-genai
python-dotenv
pydantic
ddgs
langgraph